+++ NEWS.md
@@ -1,9 +1,14 @@
 # geobounds (development version)
 
--   Add DOI.
+-   All functions:
+    -   Improve detection for Antarctica and Kosovo.
+    -   All functions return a `MULTIPOLYGON`.
+-   `get_gb*()`: In all functions now the `country` argument recognize mixed
+    types (e.g. `get_gb(country = c("Germany", "USA"))` would work).
 -   `get_gb_cgaz()` get the latest data available on the repo
     <https://github.com/wmgeolab/geoBoundaries/tree/main/releaseData>.
--   All functions: Improve detection for Antarctica and Kosovo.
+-   Add DOI.
+-   
 
 # geobounds 0.0.1